Organize EditModal components into tab folders
