+++ components/charts/ChartEditModal.tsx
@@ -4,10 +4,10 @@ import React, { useState } from "react"
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog"
 import { Button } from "@/components/ui/button"
 import { useAnalysisStore } from "@/stores/useAnalysisStore"
-import { DataSourceTab } from "./EditModal/DataSourceTab"
-import { ParametersTab } from "./EditModal/ParametersTab"
-import { AppearanceTab } from "./EditModal/AppearanceTab"
-import { ReferenceLineTab } from "./EditModal/ReferenceLineTab"
+import { DataSourceTab } from "./EditModal/data-source"
+import { ParametersTab } from "./EditModal/parameters"
+import { AppearanceTab } from "./EditModal/appearance"
+import { ReferenceLineTab } from "./EditModal/reference-lines"
 import { ChartPreview } from "./ChartPreview"
 import { EventInfo } from "@/types"
