+++ components/charts/EditModal/appearance/AppearanceTab.tsx
