+++ components/charts/EditModal/appearance/PlotStyleTable.tsx
