+++ components/charts/EditModal/appearance/TitleAndOptionsSection.tsx
