+++ components/charts/EditModal/appearance/YAxisSettings.tsx
