+++ components/charts/EditModal/appearance/index.ts
@@ -1,0 +1,6 @@
+export { AppearanceTab } from './AppearanceTab'
+export { ChartTypeSelector } from './ChartTypeSelector'
+export { TitleAndOptionsSection } from './TitleAndOptionsSection'
+export { XAxisSettings } from './XAxisSettings'
+export { YAxisSettings } from './YAxisSettings'
+export { PlotStyleTable } from './PlotStyleTable'
