+++ components/charts/EditModal/data-source/index.ts
@@ -1,0 +1,1 @@
+export { DataSourceTab } from './DataSourceTab'
