+++ components/charts/EditModal/parameters/ParametersTab.tsx
