+++ components/charts/EditModal/parameters/ThresholdColorSection.tsx
