+++ components/charts/EditModal/parameters/ThresholdPointsTable.tsx
