+++ components/charts/EditModal/parameters/XParameterSettings.tsx
