+++ components/charts/EditModal/parameters/YParametersSettings.tsx
